refactor: extract AtomicJsonFile for cross-platform atomic config writes
Replace File.Replace (destination-must-exist, platform edge cases) with
File.Move(overwrite: true) — an atomic rename on the same volume on both
Windows and Unix that also handles the create-new case. Clean up the temp
file if the write or rename throws.

Fold the duplicated LoadJsonObject/WriteConfigAtomic helpers from
DiscordConfigPersistence and ModelCatalogPersistence into a single shared
AtomicJsonFile helper.
